+++ src/main.ts
@@ -3,7 +3,7 @@ import {findTaggedFiles} from "./util/scan";
 import {Quote, updateQuotesVault} from "./types/quote";
 import {processCodeblock} from "./processor/codeblock";
 import {DEFAULT_SETTINGS, LocalQuotesSettings, LocalQuotesSettingTab} from "./settings";
-import {ErrorModal, QuoteMakerModal} from "./processor/modal";
+import {QuoteMakerModal, QuoteVaultErrorModal} from "./processor/modal";
 
 export default class LocalQuotes extends Plugin {
 	settings: LocalQuotesSettings;
@@ -37,10 +37,7 @@ export default class LocalQuotes extends Plugin {
 				if (this.quoteVault && this.quoteVault.length > 0) {
 					new QuoteMakerModal(this).open();
 				} else {
-					new ErrorModal(
-						'❌ Local Quote Error',
-						'After scan there is no any quote listings in your vault.'
-					).open();
+					new QuoteVaultErrorModal().open();
 				}
 			}
 		});